migrate to bevy 0.14

+++ src/configs.rs
@@ -5,10 +5,10 @@ pub const WH: f32 = 900.0;
 // Sprites
 pub const SPRITE_SHEET_PATH: &str = "assets.png";
 pub const SPRITE_SCALE_FACTOR: f32 = 3.0;
-pub const TILE_W: usize = 16;
-pub const TILE_H: usize = 16;
-pub const SPRITE_SHEET_W: usize = 8;
-pub const SPRITE_SHEET_H: usize = 8;
+pub const TILE_W: u32 = 16;
+pub const TILE_H: u32 = 16;
+pub const SPRITE_SHEET_W: u32 = 8;
+pub const SPRITE_SHEET_H: u32 = 8;
 
 // World
 pub const NUM_WORLD_DECORATIONS: usize = 500;
